Use prompts instead of inquirer + review init questions + merge some fs functions to fs.utils.js
