+++ cli/src/cli-helpers/askForWalletPrivateKey.js
@@ -17,7 +17,6 @@ export async function askForWalletPrivateKey({ spinner }) {
     name: 'walletPrivateKeyAnswer',
     message:
       'What is your wallet private key? (It will be used to assert that you are the owner of the app.)',
-    mask: '*',
   });
 
   try {
@@ -32,7 +31,7 @@ export async function askForWalletPrivateKey({ spinner }) {
       type: 'confirm',
       name: 'savePrivateKeyAnswer',
       message: 'Do you want to save this private key to your config?',
-      default: false,
+      initial: false,
     },
   ]);
