--- cli/src/utils/fileExists.js
@@ -1,5 +1,0 @@
-import { stat } from 'node:fs/promises';
-
-export async function fileExists(path) {
-  return !!(await stat(path).catch(() => false));
-}
